pyproject.toml
[project]
name = "govinfo-mcp-server"
version = "0.1.0"
description = "Model Context Protocol server providing LLM-friendly access to legal cases and court data through the GovInfo API v4"
authors = [{ name = "Travis-Prall", email = "smart@travisprall.com" }]
readme = "README.md"
license = { text = "" }
requires-python = ">=3.12.0"
keywords = ["mcp", "govinfo", "legal", "api"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "fastmcp>=2.8.0",
  "httpx>=0.28.1",
  "loguru>=0.7.3",
  "python-dotenv>=1.0.0",
  "anyio>=3.0.0",
  "pydantic>=2.0.0",
  "psutil>=7.0.0",
]

[project.urls]
Homepage = "https://www.travisprall.com/"
Repository = "https://github.com/Travis-Prall"
Issues = "https://github.com/Travis-Prall/issues"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  # Testing
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  # Code quality
  "mypy>=1.12.0",
  "ruff>=0.8.0",
  "pydantic>=2.0.0", # Add pydantic here as well for mypy plugin
  # Development tools
  "ipython>=8.28.0",
  "rich>=13.9.0",
  "httpx-sse>=0.4.0",
]

[tool.uv.sources]
# Add any custom package sources if needed

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers", "--disable-warnings"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
]
